6️⃣  Refactor product, category, and region entities and services
Standardized table names in entity modules to singular form and fixed relation definitions for category and region. Updated product service logic to fully implement CRUD operations for products, categories, and regions, including proper association handling. Removed temporary ProductStatus enum from DTOs and used the one from entity. Adjusted handler and middleware to match updated service signatures and type expectations. Updated README to clarify API and database design tasks.

+++ src/middlewares/auth.rs
@@ -114,7 +114,7 @@ where
 
         let user = User::find()
             .filter(
-                entity::user::Column::Id.eq(decoded.claims.sub.parse::<u64>().map_err(|_| {
+                entity::user::Column::Id.eq(decoded.claims.sub.parse::<i32>().map_err(|_| {
                     (
                         StatusCode::UNAUTHORIZED,
                         "Invalid user ID in token".to_string(),
